Move 'Mark' class into the 'evaluation' module
• Move 'Mark' class into the 'evaluation' module because it has more meaning.
• Delete the 'evaluation' module that is empty.
