+++ linnote/evaluation.py
@@ -11,8 +11,35 @@ License: Mozilla Public License, see 'LICENSE.txt' for details.
 from itertools import groupby
 from operator import attrgetter
 from pandas import read_excel
-from linnote.result import Mark
 from linnote.student import Student
+
+
+class Mark(object):
+    """Student's mark to an evaluation."""
+
+    def __init__(self, student, evaluation, score, scale=1, bonus=0):
+        """Create a new mark."""
+        super(Mark, self).__init__()
+        self.student = student
+        self.evaluation = evaluation
+        self._raw = score / scale
+        self._bonus = bonus / scale
+
+    def __repr__(self):
+        return '<Mark of {}: {}>'.format(self.student, self.value)
+
+    @property
+    def raw(self):
+        return self._raw * self.evaluation.coefficient
+
+    @property
+    def bonus(self):
+        return self._bonus * self.evaluation.coefficient
+
+    @property
+    def value(self):
+        """The processed mark, including bonus points."""
+        return (self._raw + self._bonus) * self.evaluation.coefficient
 
 
 class Evaluation(object):
